requirements.txt
Flask==2.3.3
flask-login==0.6.3
flask-mail==0.9.1
flask-wtf==1.2.1
itsdangerous==2.1.2
Werkzeug==2.3.7
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.0
openpyxl==3.1.2
fuzzywuzzy==0.18.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
